Kill warnings
git-svn-id: http://svn.research.cc.gatech.edu/kaos/evpath/trunk@14539 8ba6a8b9-845b-0410-94f0-974be2f5df47

+++ mpi_tests/bulk_cm.c
@@ -63,7 +63,7 @@ bulk_handler(CManager cm, CMConnection conn, void *vevent, void *client_data,
     message_count++;
     if (message_count == (np - 1) * msg_count) {
 	printf("All messages received\n");
-	CMCondition_signal(cm, (int)client_data);
+	CMCondition_signal(cm, (int)(long)client_data);
     }
 }
